rockrev/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
django-filter==24.3
flake8==6.0.0
djoser==2.1.0
djangorestframework-simplejwt==4.7.2
social-auth-app-django==4.0.0
django-extensions==3.2.3
Pillow==10.4.0
drf-spectacular==0.27.1
channels==4.2.0
celery==5.4.0
drf-yasg==1.21.10